Extract parser class to its own file
